+++ app/controllers/consolidations/AssociateUcrConfirmationController.scala
@@ -23,7 +23,7 @@ import models.ReturnToStartException
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import views.html.associate_ucr_confirmation
+import views.html.associateucr.associate_ucr_confirmation
 
 @Singleton
 class AssociateUcrConfirmationController @Inject()(
