+++ app/controllers/consolidations/AssociateUcrController.scala
@@ -25,7 +25,7 @@ import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.CacheRepository
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import views.html.associate_ucr
+import views.html.associateucr.associate_ucr
 
 import scala.concurrent.{ExecutionContext, Future}
