+++ app/controllers/consolidations/DisassociateUcrConfirmationController.scala
@@ -23,7 +23,7 @@ import models.ReturnToStartException
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import views.html.disassociate_ucr_confirmation
+import views.html.disassociateucr.disassociate_ucr_confirmation
 
 import scala.concurrent.ExecutionContext
