+++ app/controllers/consolidations/ShutMucrConfirmationController.scala
@@ -21,7 +21,7 @@ import javax.inject.{Inject, Singleton}
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import views.html.shut_mucr_confirmation
+import views.html.shutmucr.shut_mucr_confirmation
 
 import scala.concurrent.ExecutionContext
