+++ app/views/shutmucr/shut_mucr_confirmation.scala.html
