+++ app/views/shutmucr/shut_mucr_summary.scala.html
