--- test/unit/connectors/WiremockTestServer.scala
@@ -1,50 +1,0 @@
-/*
- * Copyright 2020 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package unit.connectors
-
-import com.github.tomakehurst.wiremock.WireMockServer
-import com.github.tomakehurst.wiremock.client.{MappingBuilder, WireMock}
-import com.github.tomakehurst.wiremock.stubbing.StubMapping
-import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, WordSpec}
-
-trait WiremockTestServer extends WordSpec with BeforeAndAfterAll with BeforeAndAfterEach {
-
-  val wireHost = "localhost"
-  val wirePort = 20001
-  val downstreamURL: String = "http://" + wireHost + ":" + wirePort
-  val wireMockServer = new WireMockServer(wirePort)
-
-  protected def stubFor(mappingBuilder: MappingBuilder): StubMapping =
-    wireMockServer.stubFor(mappingBuilder)
-
-  override protected def beforeAll(): Unit = {
-    super.beforeAll()
-    wireMockServer.start()
-    WireMock.configureFor(wireHost, wirePort)
-  }
-
-  override protected def afterAll(): Unit = {
-    wireMockServer.stop()
-    super.afterAll()
-  }
-
-  override def beforeEach(): Unit = {
-    super.beforeEach()
-    wireMockServer.resetAll()
-  }
-
-}
