+++ test/unit/controllers/consolidations/MucrOptionsControllerSpec.scala
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package unit.controllers.consolidations
+package controllers.consolidations
 
-import controllers.consolidations.{routes, MucrOptionsController}
 import forms.MucrOptions
 import forms.MucrOptions.Create
 import models.cache.AssociateUcrAnswers
@@ -32,7 +31,7 @@ import testdata.CommonTestData
 import testdata.ConsolidationTestData.validMucr
 import unit.controllers.ControllerLayerSpec
 import unit.repository.MockCache
-import views.html.mucr_options
+import views.html.associateucr.mucr_options
 
 import scala.concurrent.ExecutionContext.global
