+++ test/unit/controllers/consolidations/ShutMucrConfirmationControllerSpec.scala
@@ -22,7 +22,7 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import play.twirl.api.HtmlFormat
 import unit.controllers.ControllerLayerSpec
-import views.html.shut_mucr_confirmation
+import views.html.shutmucr.shut_mucr_confirmation
 
 import scala.concurrent.ExecutionContext.global
