+++ test/unit/views/disassociateucr/DisassociateUcrConfirmationViewSpec.scala
@@ -14,34 +14,35 @@
  * limitations under the License.
  */
 
-package views
+package views.disassociateucr
 
 import base.Injector
 import controllers.routes
-import models.cache.AssociateUcrAnswers
+import models.cache.DisassociateUcrAnswers
 import play.twirl.api.Html
 import testdata.CommonTestData.correctUcr
-import views.html.associate_ucr_confirmation
+import views.ViewSpec
+import views.html.disassociateucr.disassociate_ucr_confirmation
 import views.tags.ViewTest
 
 @ViewTest
-class AssociateDucrConfirmationViewSpec extends ViewSpec with Injector {
+class DisassociateUcrConfirmationViewSpec extends ViewSpec with Injector {
 
-  private implicit val request = journeyRequest(AssociateUcrAnswers())
-  private val page = instanceOf[associate_ucr_confirmation]
+  private implicit val request = journeyRequest(DisassociateUcrAnswers())
+  private val page = instanceOf[disassociate_ucr_confirmation]
+  private val view: Html =
+    page("DUCR", correctUcr)(request, messages)
 
-  private val view: Html = page("ducr", correctUcr)(request, messages)
-
-  "Associate Ducr Confirmation View" should {
+  "Disassociate Ucr Confirmation View" should {
 
     "have title" in {
 
-      view.getTitle must containMessage("associate.ducr.confirmation.tab.heading")
+      view.getTitle must containMessage("disassociate.ucr.confirmation.DUCR.title")
     }
 
     "have heading" in {
 
-      view.getElementsByClass("govuk-panel__title").first() must containMessage("associate.ducr.confirmation.heading", correctUcr)
+      view.getElementsByClass("govuk-panel__title").first() must containMessage("disassociate.ucr.confirmation.heading", "DUCR", correctUcr)
     }
 
     "have status information" in {
@@ -56,12 +57,11 @@ class AssociateDucrConfirmationViewSpec extends ViewSpec with Injector {
       view.getElementById("what-next").text() mustBe messages("movement.confirmation.whatNext")
     }
 
-    "have next steps section which" in {
+    "have next steps section" in {
 
       val nextSteps = view.getElementById("next-steps").getElementsByClass("govuk-link")
       nextSteps.first() must haveHref(routes.ChoiceController.startSpecificJourney(forms.Choice.AssociateUCR.value))
       nextSteps.last() must haveHref(routes.ChoiceController.startSpecificJourney(forms.Choice.ShutMUCR.value))
-
     }
 
     "display 'Back to start page' button on page" in {
@@ -71,7 +71,6 @@ class AssociateDucrConfirmationViewSpec extends ViewSpec with Injector {
       backButton.text() mustBe messages("site.backToStart")
       backButton.first() must haveHref(routes.ChoiceController.displayChoiceForm())
     }
-
   }
 
 }
